[LTM-35] [ FIX ] refactor SpeetoWinningInfoViewController to SpeetoWinningInfoView to fix button tap issue
