+++ LottoMate/LottoMate/Views/TabViews/TabBarViewController.swift
@@ -19,7 +19,7 @@ class TabBarViewController: UITabBarController {
         homeViewController.tabBarItem = homeTabIcon
         
         // MARK: Map Tab
-        let mapViewController = SpeetoWinningInfoViewController()
+        let mapViewController = WinningNumbersDetailViewController()
         let mapTabIcon = UITabBarItem(title: "지도", image: UIImage(systemName: "map"), selectedImage: UIImage(systemName: "map.fill"))
         mapViewController.tabBarItem = mapTabIcon
